Fixed a bug reported by Jean-Philippe Jodoin.
git-svn-id: http://bgslibrary.googlecode.com/svn/trunk@19 7e9cb7ec-d131-0c6a-8031-cf146dc18c44

+++ package_bgs/lb/BGModelMog.h
@@ -27,11 +27,11 @@ namespace lb_library
 {
   namespace MixtureOfGaussians
   {
-    const int NUMBERGAUSSIANS = 3;
-    const int LEARNINGRATEMOG = 0.001;
-    const int THRESHOLDMOG = 2.5;
-    const int BGTHRESHOLDMOG = 0.5;
-    const int INITIALVARMOG = 50.0;
+    const unsigned int NUMBERGAUSSIANS = 3;
+    const float LEARNINGRATEMOG = 0.001f;
+    const float THRESHOLDMOG = 2.5f;
+    const float BGTHRESHOLDMOG = 0.5f;
+    const float INITIALVARMOG = 50.0f;
 
     typedef struct tagMOGDATA
     {
